fix: Correção de Integer para Long

+++ src/main/java/com/imobiliaria/repositories/ImobiliariaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImobiliariaRepository extends JpaRepository<Imobiliaria,Integer> {
+public interface ImobiliariaRepository extends JpaRepository<Imobiliaria,Long> {
 
     boolean existsByCnpj(String cnpj);
 }
