+++ src/main/java/com/imobiliaria/repositories/ImovelRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ImovelRepository extends JpaRepository<Imovel,Integer> {
+public interface ImovelRepository extends JpaRepository<Imovel,Long> {
 
     boolean existsByCnpj(String cnpj);
